+++ scripts/ida.py
@@ -31,14 +31,31 @@ class IDA(object):
         feature toggle type in an IDA, parse out the information relevant
         to each toggle and add it to this IDA.
         """
-        with io.open(dump_file_path, 'r', encoding='utf-8') as dump_file:
-            dump_contents = json.loads(dump_file.read())
+        with io.open(dump_file_path) as dump_file:
+            try:
+                dump_contents = json.loads(dump_file.read())
+            except:
+                LOGGER.error(
+                'Loading json file at: {} failed, check toogle data in file is formated correctly'.format(dump_file_path)
+                )
+                raise
+        self._add_toggle_data(dump_contents)
+
+    def _add_toggle_data(self, dump_contents):
         for row in dump_contents:
             toggle_name = row['fields']['name']
-            toggle_type = row['model']
+            # convert sql dump model name to annotation report toggle type name
+            if row['model'] == "waffle.flag":
+                toggle_type = "WaffleFlag"
+            elif row['model'] == "waffle.switch":
+                toggle_type = "WaffleSwitch"
+            else:
+                toggle_type = row['model']
+
             toggle_data = row['fields']
             toggle_state = ToggleState(toggle_type, toggle_data)
             toggle = Toggle(toggle_name, toggle_state)
+
             self.toggles[toggle_type].append(toggle)
         LOGGER.info(
             'Finished collecting toggle state for {}'.format(self.name)
@@ -194,11 +211,13 @@ def add_toggle_state_to_idas(idas, dump_file_path):
     """
     ida_name_pattern = re.compile(r'(?P<ida>[a-z]*)_.*json')
     sql_dump_files = [
-        f for f in os.listdir(dump_file_path) if re.search(ida_name_pattern, f)
+        f for f in os.listdir(dump_file_path) if ida_name_pattern.search(f)
     ]
     for sql_dump_file in sql_dump_files:
         sql_dump_file_path = os.path.join(dump_file_path, sql_dump_file)
-        ida_name = re.search(ida_name_pattern, sql_dump_file).group('ida')
+        ida_name = ida_name_pattern.search(sql_dump_file).group('ida')
+        if ida_name not in idas:
+            idas[ida_name] = IDA(ida_name)
         LOGGER.info(
             'Collecting toggle_state from {} for {}'.format(
                 sql_dump_file_path, ida_name
@@ -215,16 +234,18 @@ def add_toggle_annotations_to_idas(idas, annotation_report_files_path):
     each file, parsing and linking the annotation data to the toggle state
     data in the IDA.
     """
-    ida_name_pattern = re.compile(r'(?P<ida>[a-z]*)-annotations.yml')
+    ida_name_pattern = re.compile(r'(?P<ida>[a-z]*)_annotations.yml')
     annotation_files = [
         f for f in os.listdir(annotation_report_files_path)
-        if re.search(ida_name_pattern, f)
+        if ida_name_pattern.search(f)
     ]
     for annotation_file in annotation_files:
         annotation_file_path = os.path.join(
             annotation_report_files_path, annotation_file
         )
-        ida_name = re.search(ida_name_pattern, annotation_file).group('ida')
+        ida_name = ida_name_pattern.search(annotation_file).group('ida')
+        if ida_name not in idas:
+            idas[ida_name] = IDA(ida_name)
         LOGGER.info(
             'Collecting annotations from {} for {}'.format(
                 annotation_file, ida_name